Moved App to folder and changed port to 3000
